UnusedMethodElimination fix + x64 warning fixes
With UnusedMethodElimination, if an object has no methods called, then
the cognew/coginit referenced methods will not be marked as called since
the methods containeding the coginit/cognew are not called. This fixes
something mentioned in issue #21
Added x64 compile options in the VS2017 project files for the
PropellerCompiler.lib, and fixed up a bunch of warnings related to
compiling in x64 mode. Note: openspin itself is still x86 (32bit) and
still compiles the lib in x86(32bit) mode.

+++ PropellerCompiler/CompileSpin.cpp
@@ -64,7 +64,7 @@ static bool GetPASCIISource(char* pFilename)
             pp_push_file_struct(&s_preprocessor, &mfile, pFilename);
             pp_run(&s_preprocessor);
             pBuffer = pp_finish(&s_preprocessor);
-            nLength = strlen(pBuffer);
+            nLength = (int)strlen(pBuffer);
             if (nLength == 0)
             {
                 free(pBuffer);
@@ -523,7 +523,7 @@ static void DumpSymbols()
 
 static void DumpList()
 {
-    int listOffset = 0;
+    size_t listOffset = 0;
     while (listOffset < s_pCompilerData->list_length)
     {
         char* pTemp = strstr(&(s_pCompilerData->list[listOffset]), "\r");
@@ -546,7 +546,7 @@ static void DumpList()
 
 static void DumpDoc()
 {
-    int docOffset = 0;
+    size_t docOffset = 0;
     while (docOffset < s_pCompilerData->doc_length)
     {
         char* pTemp = strstr(&(s_pCompilerData->doc[docOffset]), "\r");
